Add tests for bind function
